# Fix BLS performance